backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.6
redis==5.0.1
django-redis==5.4.0

# Environment Variables
python-decouple==3.8

# Security & Threat Intelligence
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# MITRE ATT&CK Data
stix2==3.0.1

# API Documentation
drf-yasg==1.21.7

# Testing
pytest==7.4.4
pytest-django==4.7.0
faker==22.0.0

# Code Quality
black==24.1.1
flake8==7.0.0

# Utilities
pytz==2023.3
gunicorn==21.2.0